fix: generate multiple intervals instead of separate issues
